+++ src/test/2024-04/NGFS_exp.sol
@@ -23,24 +23,30 @@ interface IPancakeFactory {
 
 interface IPancakeRouter {
     function swapExactTokensForTokensSupportingFeeOnTransferTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external;
 }
 
 interface INGFSToken {
     function delegateCallReserves() external;
-    function setProxySync(address) external;
-    function balanceOf(address) external view returns (uint256);
+    function setProxySync(
+        address
+    ) external;
+    function balanceOf(
+        address
+    ) external view returns (uint256);
     function reserveMultiSync(address, uint256) external;
     function approve(address, uint256) external returns (bool);
 }
 
 interface IBEP20 {
-    function balanceOf(address) external view returns (uint256);
+    function balanceOf(
+        address
+    ) external view returns (uint256);
 }
 
 contract NGFS is Test {
@@ -74,11 +80,7 @@ contract NGFS is Test {
 
         uint256 deadline = 1_714_043_885;
         IPancakeRouter(PANCAKE_ROUTER).swapExactTokensForTokensSupportingFeeOnTransferTokens(
-            amount,
-            0,
-            path,
-            address(this),
-            deadline
+            amount, 0, path, address(this), deadline
         );
 
         uint256 tokenBalanceAfter = IBEP20(USDT_TOKEN).balanceOf(address(this));
